Rename "text-input-v2" to "text-input"
