+++ catalog/input/demos/inferred.jsx
@@ -1,6 +1,6 @@
 import React, { Component, createRef } from 'react';
 import styled from 'styled-components';
-import { InferredSelect, InferredText, Select } from '../../../components/text-input-v2';
+import { InferredSelect, InferredText, Select } from '../../../components/text-input';
 import { Button } from '../../../index';
 
 const DemoContainer = styled.div`
